requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai==1.54.0
requests==2.32.3
